move serialization test
